+++ lib/Castor/Castor.php
@@ -5,7 +5,7 @@
  * @contributor Laurent Jouanneau
  *
  * @copyright   2006 Loic Mathaud
- * @copyright   2006-2020 Laurent Jouanneau
+ * @copyright   2006-2025 Laurent Jouanneau
  *
  * @link        http://www.jelix.org
  * @licence     GNU Lesser General Public Licence see LICENCE file or http://www.gnu.org/licenses/lgpl.html
@@ -37,7 +37,7 @@ class Castor extends CastorCore
      * include the compiled template file and call one of the generated function.
      *
      * @param string $tpl        template selector
-     * @param string $outputType the type of output (html, text etc..)
+     * @param string $outputType the type of output (html, text etc..) (deprecated)
      * @param bool   $trusted    says if the template file is trusted or not
      *
      * @return string the suffix name of the function to call
@@ -59,7 +59,17 @@ class Castor extends CastorCore
             throw new Exception('cache path is invalid ! its value is: "'.$this->config->cachePath.'".');
         }
 
-        $cachefile = $this->config->cachePath.$cachefile . $outputType . ($trusted ? '_t' : '') . '_' . basename($tpl);
+        if ($outputType == 'html') {
+            // no prefix for default outputType and trust level.
+            // outputType and trust level may be defined with pragma instruction
+            // so the prefix may not correspond to the real output / trust level
+            $prefixFileName = ($trusted ? '' : 'nt_');
+        }
+        else {
+            $prefixFileName = $outputType.($trusted ? '' : '_nt').'_';
+        }
+
+        $cachefile = $this->config->cachePath.$cachefile . $prefixFileName . basename($tpl).'.php';
 
         $mustCompile = $this->config->compilationForce || !file_exists($cachefile);
         if (!$mustCompile) {
@@ -70,15 +80,26 @@ class Castor extends CastorCore
 
         if ($mustCompile) {
             $compiler = $this->getCompiler();
+            if (preg_match('/\\.ctpl$/', $tpl)) {
+                $compiler->setSyntaxVersion(2);
+            }
             $compiler->compile($this->_templateName,
                                $tpl, $outputType, $trusted,
                                $this->userModifiers, $this->userFunctions);
         }
         require_once $cachefile;
 
-        return md5($tpl.'_' . $outputType . ($trusted ? '_t' : ''));
+        return md5($prefixFileName.$tpl);
     }
 
+    /**
+     * @param string $tpl template selector
+     * @param string $outputType the type of output (html, text etc..) (deprecated)
+     * @param bool $trusted says if the template file is trusted or not
+     * @param bool $callMeta
+     * @return false|string
+     * @throws \Exception
+     */
     public function fetch($tpl, $outputType = '', $trusted = true, $callMeta = true)
     {
         return $this->_fetch($tpl, $tpl, $outputType, $trusted, $callMeta);