+++ lib/Castor/CastorCore.php
@@ -4,7 +4,7 @@
  * @author      Laurent Jouanneau
  * @contributor Dominique Papin
  *
- * @copyright   2005-2022 Laurent Jouanneau, 2007 Dominique Papin
+ * @copyright   2005-2025 Laurent Jouanneau, 2007 Dominique Papin
  *
  * @link        http://www.jelix.org
  * @licence     GNU Lesser General Public Licence see LICENCE file or http://www.gnu.org/licenses/lgpl.html
@@ -173,7 +173,7 @@ abstract class CastorCore
      * process all meta instruction of a template.
      *
      * @param string $tpl        template selector
-     * @param string $outputType the type of output (html, text etc..)
+     * @param string $outputType the type of output (html, text etc..) (deprecated)
      * @param bool   $trusted    says if the template file is trusted or not
      */
     public function meta($tpl, $outputType = '', $trusted = true)
@@ -197,7 +197,7 @@ abstract class CastorCore
      * display the generated content from the given template.
      *
      * @param string $tpl        template selector
-     * @param string $outputType the type of output (html, text etc..)
+     * @param string $outputType the type of output (html, text etc..) (deprecated)
      * @param bool   $trusted    says if the template file is trusted or not
      */
     public function display($tpl, $outputType = '', $trusted = true)
@@ -237,7 +237,7 @@ abstract class CastorCore
      * include the compiled template file and call one of the generated function.
      *
      * @param string $tpl        template selector
-     * @param string $outputType the type of output (html, text etc..)
+     * @param string $outputType the type of output (html, text etc..) (deprecated)
      * @param bool   $trusted    says if the template file is trusted or not
      *
      * @return string the suffix name of the function to call
@@ -248,7 +248,7 @@ abstract class CastorCore
      * return the generated content from the given template.
      *
      * @param string $tpl        template selector
-     * @param string $outputType the type of output (html, text etc..)
+     * @param string $outputType the type of output (html, text etc..) (deprecated)
      * @param bool   $trusted    says if the template file is trusted or not
      * @param bool   $callMeta   false if meta should not be called
      *
@@ -259,7 +259,7 @@ abstract class CastorCore
     /**
      * @param string $tpl        the template name
      * @param string $getTemplateArg
-     * @param string $outputType the type of output (html, text etc..)
+     * @param string $outputType the type of output (html, text etc..) (deprecated)
      * @param bool   $trusted    says if the template file is trusted or not
      * @param bool   $callMeta   false if meta should not be called
      *
@@ -302,28 +302,37 @@ abstract class CastorCore
     /**
      * Return the generated content from the given string template (virtual).
      *
-     * @param string $tpl        template content
-     * @param string $outputType the type of output (html, text etc..)
+     * @param string $templateContent        template content
+     * @param string $outputType the type of output (html, text etc..) (deprecated)
      * @param bool   $trusted    says if the template file is trusted or not
      * @param bool   $callMeta   false if meta should not be called
      *
      * @return string the generated content
      */
-    public function fetchFromString($tpl, $outputType = '', $trusted = true, $callMeta = true)
+    public function fetchFromString($templateContent, $outputType = '', $trusted = true, $callMeta = true)
     {
         ob_start();
         try {
-            $cachePath = $this->getCachePath().'virtual/';
-
             $previousTpl = $this->_templateName;
-            $md = 'virtual_'.md5($tpl).($trusted ? '_t' : '');
-            $this->_templateName = $md;
 
             if ($outputType == '') {
                 $outputType = 'html';
             }
 
-            $cachePath .= $outputType . '_' . $this->_templateName . '.php';
+            if ($outputType == 'html') {
+                // no prefix for default outputType and trust level.
+                // outputType and trust level may be defined with pragma instruction
+                // so the prefix may not correspond to the real output / trust level
+                $prefixFileName = ($trusted ? '' : 'nt_');
+            }
+            else {
+                $prefixFileName = $outputType.($trusted ? '' : '_nt').'_';
+            }
+
+            $md = 'virtual_'.md5($prefixFileName.$templateContent);
+            $this->_templateName = $md;
+            $cachePath = $this->getCachePath().'virtual/';
+            $cachePath .= $prefixFileName . $this->_templateName . '.php';
 
             $mustCompile = $this->compilationNeeded($cachePath);
 
@@ -331,7 +340,7 @@ abstract class CastorCore
                 $compiler = $this->getCompiler();
                 $compiler->outputType = $outputType;
                 $compiler->trusted = $trusted;
-                $compiler->compileString($tpl,
+                $compiler->compileString($templateContent,
                                          $cachePath,
                                          $this->userModifiers,
                                          $this->userFunctions,
@@ -382,7 +391,7 @@ abstract class CastorCore
      * register a user function. The function should accept at least a CastorCore object
      * as first parameter.
      *
-     * @param string $name         the name of the modifier in a template
+     * @param string $name         the name of the function in a template
      * @param string $functionName the corresponding PHP function
      */
     public function registerFunction($name, $functionName)