+++ lib/Castor/CompilerCore.php
@@ -133,6 +133,8 @@ abstract class CompilerCore
 
     protected $_autoescape = false;
 
+    protected $_syntaxVersion = 1;
+
     /**
      * list of user functions.
      */
@@ -140,6 +142,7 @@ abstract class CompilerCore
 
     protected $removeASPtags = true;
 
+    protected $generatedContentStarted = false;
     /**
      * Initialize some properties.
      */
@@ -179,14 +182,61 @@ abstract class CompilerCore
     }
 
 
-    public function compileString($templateContent, $cacheFile, $userModifiers, $userFunctions, $md5, $header = '', $footer = '')
+    public function setSyntaxVersion($syntaxVersion)
     {
+        $this->_syntaxVersion = $syntaxVersion;
+    }
+
+    public function getSyntaxVersion()
+    {
+        return $this->_syntaxVersion;
+    }
+
+    /**
+     * Compile a template and save the result into the given file
+     *
+     * @param string  $templateContent  the template content
+     * @param string  $cacheFile        the filename where to store the result
+     * @param array   $userModifiers    list of modifiers to declare (deprecated)
+     * @param array   $userFunctions    list of template function to declare (deprecated)
+     * @param string  $uniqIdentifier   identifier of the template. Used as suffix of generated PHP functions
+     * @param string  $header           PHP content to add as header of the generated template
+     * @param string  $footer           PHP content to add as footer of the generated template
+     */
+    public function compileString($templateContent, $cacheFile, $userModifiers, $userFunctions, $uniqIdentifier, $header = '', $footer = '')
+    {
+
+        $result = $this->_compileString($templateContent, $userModifiers, $userFunctions, $uniqIdentifier);
+
+        $result = "<?php\n".$header.$result.$footer;
+
+        $this->_saveCompiledString($cacheFile, $result);
+
+        return true;
+    }
+
+    /**
+     * Compile a template: generate the full PHP code
+     *
+     * @param string  $templateContent  the template content
+     * @param string  $cacheFile        the filename where to store the result
+     * @param array   $userModifiers    list of modifiers to declare (deprecated)
+     * @param array   $userFunctions    list of template function to declare (deprecated)
+     * @param string  $uniqIdentifier   identifier of the template. Used as suffix of generated PHP functions
+     * @param string  $header           PHP content to add as header of the generated template
+     * @param string  $footer           PHP content to add as footer of the generated template
+     *
+     * @return string the PHP code of the template, without php start tag and php end tag
+     */
+    protected function _compileString($templateContent, $userModifiers, $userFunctions, $md5)
+    {
+        $this->generatedContentStarted = false;
         $this->_modifier = array_merge($this->_modifier, $userModifiers);
         $this->_userFunctions = $userFunctions;
 
         $result = $this->compileContent($templateContent);
 
-        $header = "<?php \n".$header;
+        $header = '';
         foreach ($this->_pluginPath as $path => $ok) {
             $header .= ' require_once(\''.$path."');\n";
         }
@@ -194,11 +244,9 @@ abstract class CompilerCore
         $header .= "\n".$this->_metaBody."\n}\n";
 
         $header .= 'function template_'.$md5.'($t){'."\n?>";
-        $result = $header.$result."<?php \n}\n".$footer;
+        $result = $header.$result."<?php \n}\n";
 
-        $this->_saveCompiledString($cacheFile, $result);
-
-        return true;
+        return $result;
     }
 
     abstract protected function _saveCompiledString($cacheFile, $result);
@@ -214,6 +262,7 @@ abstract class CompilerCore
         $tplContent = preg_replace("!{\*(.*?)\*}!s", '', $tplContent);
         $tplContent = preg_replace("!{#(.*?)#}!s", '', $tplContent);
 
+        // output content `<? ... ? >` as is...
         $tplContent = preg_replace_callback("!(<\?.*\?>)!sm", function ($matches) {
             return '<?php echo \''.str_replace("'", "\\'", $matches[1]).'\'?>';
         }, $tplContent);
@@ -223,15 +272,18 @@ abstract class CompilerCore
           $tplContent = preg_replace('!<%.*%>!s', '', $tplContent);
         }
 
-        preg_match_all('!{literal}(.*?){/literal}!s', $tplContent, $_match);
-        $this->_literals = $_match[1];
-        $tplContent = preg_replace('!{literal}(.*?){/literal}!s', '{literal}', $tplContent);
+        if ($this->_syntaxVersion == 1) {
+            // replace literal content by a specific tag, that will be replaced by original content
+            preg_match_all('!{literal}(.*?){/literal}!s', $tplContent, $_match);
+            $this->_literals = $_match[1];
+            $tplContent = preg_replace('!{literal}(.*?){/literal}!s', '{literal}', $tplContent);
 
-        preg_match_all('!{verbatim}(.*?){/verbatim}!s', $tplContent, $_match);
-        $this->_verbatims = $_match[1];
-        $tplContent = preg_replace('!{verbatim}(.*?){/verbatim}!s', '{verbatim}', $tplContent);
-
-        $tplContent = preg_replace_callback("/{((.).*?)}(\n)/sm", function ($matches) {
+            // replace verbatim content (which is equivalent to literal) by a specific tag, that will be replaced by original content
+            preg_match_all('!{verbatim}(.*?){/verbatim}!s', $tplContent, $_match);
+            $this->_verbatims = $_match[1];
+            $tplContent = preg_replace('!{verbatim}(.*?){/verbatim}!s', '{verbatim}', $tplContent);
+            // remove \n after all tags except for variable output instructions
+            $tplContent = preg_replace_callback("/{((.).*?)}(\n)/sm", function ($matches) {
                 list($full, , $firstCar, $lastcar) = $matches;
                 if ($firstCar == '=' || $firstCar == '$' || $firstCar == '@') {
                     return "$full\n";
@@ -239,8 +291,33 @@ abstract class CompilerCore
                     return $full;
                 }
             }, $tplContent);
-        $tplContent = preg_replace_callback('/{((.).*?)}/sm', array($this, '_callback'), $tplContent);
 
+            // process every template instruction
+            $tplContent = preg_replace_callback('/{((.).*?)}/sm', array($this, '_callbackV1'), $tplContent);
+        }
+        else {
+            // replace literal content by a specific tag, that will be replaced by original content
+            preg_match_all('!\\{\\%\s*literal\s*\\%\\}(.*?)\\{\\%\s*/literal\s*\\%\\}!s', $tplContent, $_match);
+            $this->_literals = $_match[1];
+            $tplContent = preg_replace('!\\{\\%\s*literal\s*\\%\\}(.*?)\\{\\%\s*/literal\s*\\%\\}!s', '{%literal%}', $tplContent);
+
+            // replace verbatim content (which is equivalent to literal) by a specific tag, that will be replaced by original content
+            preg_match_all('!\\{\\%\s*verbatim\s*\\%\\}(.*?)\\{\\%\s*/verbatim\s*\\%\\}!s', $tplContent, $_match);
+            $this->_verbatims = $_match[1];
+            $tplContent = preg_replace('!\\{\\%\s*verbatim\s*\\%\\}(.*?)\\{\\%\s*/verbatim\s*\\%\\}!s', '{%verbatim%}', $tplContent);
+
+            // remove \n after all tags except for variable output instructions
+            $tplContent = preg_replace_callback("/\\{[\\!\\%](.+)[\\!\\%]\\}(\n)/Usm", function ($matches) {
+                return $matches[0];
+            }, $tplContent);
+
+            // process every template instruction
+            $tplContent = preg_replace_callback('/\\{(\\!.+\\!)\\}/Usm', array($this, '_callbackV2Pragma'), $tplContent);
+            $tplContent = preg_replace_callback('/\\{\\{(.+)\\}\\}/Usm', array($this, '_callbackV2OutputInstruction'), $tplContent);
+            $tplContent = preg_replace_callback('/\\{\\%(.+)\\%\\}/Usm', array($this, '_callbackV2Instruction'), $tplContent);
+        }
+
+        // remove empty php tags
         $tplContent = preg_replace('/<\?php\\s+\?>/', '', $tplContent);
 
         if (count($this->_blockStack)) {
@@ -258,7 +335,7 @@ abstract class CompilerCore
      *
      * @return string the corresponding php code of the tag (with php tag).
      */
-    public function _callback($matches)
+    public function _callbackV1($matches)
     {
         list(, $tag, $firstcar) = $matches;
 
@@ -269,12 +346,15 @@ abstract class CompilerCore
 
         $this->_currentTag = $tag;
         if ($firstcar == '=') {
+            $this->generatedContentStarted = true;
             return  '<?php echo '.$this->_parseVariable(substr($tag, 1)).'; ?>';
         } elseif ($firstcar == '$' || $firstcar == '@') {
+            $this->generatedContentStarted = true;
             return '<?php echo ' . $this->_parseVariable($tag) . '; ?>';
         } elseif ($firstcar == '!') {
             return $this->_parsePragma($tag);
         } else {
+            $this->generatedContentStarted = true;
             if (!preg_match('/^(\/?[a-zA-Z0-9_]+)(?:(?:\s+(.*))|(?:\((.*)\)))?$/ms', $tag, $m)) {
                 $this->doError1('errors.tpl.tag.function.invalid', $tag);
             }
@@ -291,8 +371,41 @@ abstract class CompilerCore
                 return '}';
             }
 
-            return '<?php '.$this->_parseFunction($m[1], $m[2]).'?>';
+            return '<?php '.$this->_parseInstruction($m[1], $m[2]).'?>';
         }
+    }
+
+    public function _callbackV2Pragma($matches)
+    {
+        $tag = $matches[1];
+        return $this->_parsePragma($tag);
+    }
+
+    public function _callbackV2OutputInstruction($matches)
+    {
+
+        $tag = trim($matches[1]);
+        $this->generatedContentStarted = true;
+        $this->_currentTag = $tag;
+        return '<?php echo ' . $this->_parseVariable($tag) . '; ?>';
+    }
+
+    public function _callbackV2Instruction($matches)
+    {
+        $tag = trim($matches[1]);
+        $this->_currentTag = $tag;
+        $this->generatedContentStarted = true;
+        if (!preg_match('/^(\/?[a-zA-Z0-9_]+)(?:(?:\s+(.*))|(?:\((.*)\)))?$/ms', $tag, $m)) {
+            $this->doError1('errors.tpl.tag.function.invalid', $tag);
+        }
+        if (count($m) == 4) {
+            $m[2] = $m[3];
+        }
+        if (!isset($m[2])) {
+            $m[2] = '';
+        }
+
+        return '<?php '.$this->_parseInstruction($m[1], $m[2]).'?>';
     }
 
     /**
@@ -304,7 +417,7 @@ abstract class CompilerCore
      */
     protected function _parseVariable($expr)
     {
-        $tok = explode('|', $expr);
+        $tok = preg_split('/\s*\\|\s*/', $expr);
         $res = $this->_parseFinal(array_shift($tok), $this->_allowedInVar, $this->_excludedInVar);
 
         $hasEscHtmlModifier = false;
@@ -362,14 +475,14 @@ abstract class CompilerCore
     }
 
     /**
-     * analyse the tag which have a name.
+     * analyse the tag that have a name.
      *
      * @param string $name the name of the tag
      * @param string $args the content that follow the name in the tag
      *
      * @return string the corresponding php instructions
      */
-    protected function _parseFunction($name, $args)
+    protected function _parseInstruction($name, $args)
     {
         $res = '';
         switch ($name) {
@@ -433,6 +546,10 @@ abstract class CompilerCore
             case '/for':
             case '/if':
             case '/while':
+            case 'endforeach':
+            case 'endfor':
+            case 'endif':
+            case 'endwhile':
                 $short = substr($name, 1);
                 if (end($this->_blockStack) != $short) {
                     $this->doError1('errors.tpl.tag.block.end.missing', end($this->_blockStack));
@@ -466,6 +583,8 @@ abstract class CompilerCore
 
             case '/literal':
             case '/verbatim':
+            case 'endliteral':
+            case 'endverbatim':
                 $this->doError1('errors.tpl.tag.block.begin.missing', substr($name, 1));
                 break;
 
@@ -497,6 +616,7 @@ abstract class CompilerCore
                 break;
 
             case '/meta_if':
+            case 'meta_endif':
                 $short = substr($name, 1);
                 if (end($this->_blockStack) != $short) {
                     $this->doError1('errors.tpl.tag.block.end.missing', end($this->_blockStack));
@@ -507,9 +627,10 @@ abstract class CompilerCore
                 break;
 
             default:
-                if (preg_match('!^/(\w+)$!', $name, $m)) {
-                    if (end($this->_blockStack) != $m[1]) {
-                        $this->doError1('errors.tpl.tag.block.end.missing', end($this->_blockStack));
+                if (preg_match('!^(?:end|/)(\w+)$!', $name, $m)) {
+                    $currentBlock = end($this->_blockStack);
+                    if ($currentBlock != $m[1]) {
+                        $this->doError1('errors.tpl.tag.block.end.missing', $currentBlock);
                     } else {
                         array_pop($this->_blockStack);
                         if (function_exists($fct = 'jtpl_block_'.$this->outputType.'_'.$m[1])) {
@@ -534,25 +655,34 @@ abstract class CompilerCore
                     $fct = $path[1];
                     $res = $fct($this, true, $argfct);
                     array_push($this->_blockStack, $name);
-                } elseif ($path = $this->_getPlugin('cfunction', $name)) {
-                    require_once $path[0];
-                    $argfct = $this->_parseFinal($args, $this->_allowedAssign, array(';'), true);
-                    $fct = $path[1];
-                    $res = $fct($this, $argfct);
-                } elseif ($path = $this->_getPlugin('function', $name)) {
-                    $argfct = $this->_parseFinal($args, $this->_allowedAssign);
-                    $res = $path[1].'( $t'.(trim($argfct) != '' ? ','.$argfct : '').');';
-                    $this->_pluginPath[$path[0]] = true;
-                } elseif (isset($this->_userFunctions[$name])) {
-                    $argfct = $this->_parseFinal($args, $this->_allowedAssign);
-                    $res = $this->_userFunctions[$name].'( $t'.(trim($argfct) != '' ? ','.$argfct : '').');';
                 } else {
-                    $this->doError1('errors.tpl.tag.function.unknown', $name);
+                    $res = $this->_parseFunction($name, $args);
                 }
         }
 
         return $res;
     }
+
+    protected function _parseFunction($name, $args)
+    {
+        if ($path = $this->_getPlugin('cfunction', $name)) {
+            require_once $path[0];
+            $argfct = $this->_parseFinal($args, $this->_allowedAssign, array(';'), true);
+            $fct = $path[1];
+            $res = $fct($this, $argfct);
+        } elseif ($path = $this->_getPlugin('function', $name)) {
+            $argfct = $this->_parseFinal($args, $this->_allowedAssign);
+            $res = $path[1].'( $t'.(trim($argfct) != '' ? ','.$argfct : '').');';
+            $this->_pluginPath[$path[0]] = true;
+        } elseif (isset($this->_userFunctions[$name])) {
+            $argfct = $this->_parseFinal($args, $this->_allowedAssign);
+            $res = $this->_userFunctions[$name].'( $t'.(trim($argfct) != '' ? ','.$argfct : '').');';
+        } else {
+            $this->doError1('errors.tpl.tag.function.unknown', $name);
+        }
+        return $res;
+    }
+
 
     /**
      * analyse a pragma tag : `{! !}`
@@ -561,21 +691,69 @@ abstract class CompilerCore
      */
     protected function _parsePragma($expr)
     {
-        if (!preg_match('/^\!\s*([a-z0-9_]+)\s*(?:=\s*([a-z0-9_]+)\s*)?\!$/', $expr, $m)) {
+        if (!preg_match('/^\!\s*([a-z0-9_\-]+)\s*(?:=\s*([a-z0-9_]+)\s*)?\!$/', $expr, $m)) {
             $this->doError1('errors.tpl.tag.syntax.invalid', '{'.$expr.'}');
         }
-
-        switch($m[1]) {
+        $parameter = $m[1];
+        $value = $m[2] ?? null;
+        switch($parameter) {
             case 'autoescape':
-                if (isset($m[2])) {
-                    $this->_autoescape = in_array(strtolower($m[2]), array('true', 'on', '1'));
+                if ($value !== null) {
+                    $this->_autoescape = in_array(strtolower($value), array('true', 'on', '1'));
                 }
                 else {
                     $this->_autoescape = true;
                 }
-                return '';
+                break;
+            case 'output-type':
+                if ($this->generatedContentStarted) {
+                    $this->doError1('errors.tpl.tag.pragma.too.late', '{'.$expr.'}');
+                }
+                if ($value !== null) {
+                    $this->outputType = $value;
+                }
+                else {
+                    $this->doError1('errors.tpl.tag.pragma.missing.value', '{'.$expr.'}');
+                }
+                break;
+            case 'trusted':
+                if ($this->generatedContentStarted) {
+                    $this->doError1('errors.tpl.tag.pragma.too.late', '{'.$expr.'}');
+                }
+                if ($value !== null) {
+                    // "trusted" value given to the api has priority over the pragma instruction, because of potential
+                    // security issue. In case of untrusted template, we cannot trust the pragma instruction.
+                    // So if the API indicates an untrusted template, we should ignore the pragma instruction, else
+                    // this instruction may set "trusted" as true whereas the template must be considered as untrusted.
+                    if ($this->trusted === true) {
+                        $this->trusted = in_array(strtolower($value), array('true', 'on', '1'));
+                    }
+                }
+                else {
+                    $this->trusted = true;
+                }
+                break;
+/*            case 'syntax':
+                if ($this->generatedContentStarted) {
+                    $this->doError1('errors.tpl.tag.pragma.too.late', '{'.$expr.'}');
+                }
+                if ($value !== null) {
+                    $value = intval($value);
+                    if ($value == 1 || $value == 2) {
+                        $this->_syntaxVersion = $value;
+                    }
+                    else {
+                        $this->doError1('errors.tpl.tag.pragma.bad.value', '{'.$expr.'}');
+                    }
+                }
+                else {
+                    $this->doError1('errors.tpl.tag.pragma.missing.value', '{'.$expr.'}');
+                }
+                break;*/
+            default:
+                $this->doError1('errors.tpl.tag.pragma.unknown', '{'.$expr.'}');
         }
-        $this->doError1('errors.tpl.tag.pragma.unknown', '{'.$expr.'}');
+        return '';
     }
 
     /**