+++ tests/exemple.php
@@ -14,7 +14,7 @@ $tpl->assign('countries', $countries);
 $tpl->assign('titre', 'This is a test !');
 $tpl->display('test.tpl');
 
-$tpl = new \Jelix\Castor\Castor();
+$tpl = new \Jelix\Castor\Castor($config);
 $tpl->assign('titre', 'This is an other test !');
 $tpl->display('foo/test.tpl');