+++ CHANGELOG.md
@@ -4,6 +4,17 @@ Changelog
 next
 -----
 
+
+1.2.0
+-----
+
+- New syntax for templates instructions, which is available for templates files having the `.ctpl` extension.
+- New way to indicate the output type. You can indicate the output type directly into the template with a pragma
+  instruction: `{! output-type = text !}`. It overrides the `$outputType` parameter given to Castor.
+- New way to indicate the level of trust: support of the pragma instruction `{! trusted = false !}` in templates.
+  However, the `$trusted` parameter has priority over the value of this instruction to avoid security issues.
+- The `$outputType` parameter of methods `Castor::fetch()`, `Castor::display()`, `Castor::meta()`, `CastorCore::getTemplate()`, `Compiler::compile()`, is deprecated.
+
 1.1.5
 -----